Rationalize handling of have_foo for foo in {exif jpg tiff zlib}.  For each, set have_foo to yes if the library is found, and otherwise no. Solves problem of have_foo being "" if the header check failed (e.g. in have_tiff), causing a shell syntax error.
Improve comment about pow and libm.

For tiff check, set have_tiff, not have_zlib.

Explain the reason for the odd (to an autoconf user) -DNO_JPEG usage.

Only print a message about TIFF compression support if we have tiff to
start with; it's odd to report having TIFF lossless compression but
not tiff.  Use the word 'deflate', so the clueful will realize we mean
that and not LZW or packbits.